+++ accountmonth/commands.go
@@ -52,7 +52,7 @@ type RegisterNewAccountMovement struct {
 	AccountMovementId AccountMovementId     `json:"account_movement_id"`
 	MovementTypeId    *movementtype.Id      `json:"movement_type_id" eh:"optional"`
 	Action            common.MovementAction `json:"action"`
-	Amount            float64               `json:"amount"`
+	Amount            float32               `json:"amount"`
 	Date              time.Time             `json:"date"`
 	SourceAccountId   *account.Id           `json:"source_account_id" eh:"optional"`
 	Description       string                `json:"description"`
@@ -75,7 +75,7 @@ func (command RegisterNewAccountMovement) CommandType() eventhorizon.CommandType
 type StartAccountMonth struct {
 	AccountMonthId Id         `json:"account_month"`
 	AccountId      account.Id `json:"account_id"`
-	StartBalance   float64    `json:"start_balance"`
+	StartBalance   float32    `json:"start_balance"`
 	Month          time.Month `json:"month"`
 	Year           uint       `json:"year"`
 }
@@ -95,7 +95,7 @@ func (command StartAccountMonth) CommandType() eventhorizon.CommandType {
 type EndAccountMonth struct {
 	AccountMonthId Id         `json:"account_month"`
 	AccountId      account.Id `json:"account_id"`
-	EndBalance     float64    `json:"end_balance"`
+	EndBalance     float32    `json:"end_balance"`
 	Month          time.Month `json:"month"`
 	Year           uint       `json:"year"`
 }
