+++ accountmonth/readmodelrepository.go
@@ -19,7 +19,7 @@ type ReadModelWriter interface {
 		ctx context.Context,
 		accountMonthId *Id,
 		accountId *account.Id,
-		startBalance float64,
+		startBalance float32,
 		month time.Month,
 		year uint,
 	) error
@@ -56,7 +56,7 @@ func (repository *ReadModelRepository) StartMonth(
 	ctx context.Context,
 	accountMonthId *Id,
 	accountId *account.Id,
-	startBalance float64,
+	startBalance float32,
 	month time.Month,
 	year uint,
 ) error {
